[refactor/#7] Auth/User 분리
- ApiService, RemoteDataSource, Repository, DTO
