+++ app/src/main/java/org/sopt/at/data/remote/AuthRemoteDataSource.kt
@@ -1,0 +1,17 @@
+package org.sopt.at.data.remote
+
+import org.sopt.at.domain.model.User
+import org.sopt.at.remote.base.BaseResponse
+import org.sopt.at.remote.model.SignInResponse
+
+interface AuthRemoteDataSource {
+
+    suspend fun postSignUp(
+        userInfo: User
+    ): BaseResponse
+
+    suspend fun postSignIn(
+        id: String,
+        password: String
+    ): SignInResponse
+}
