+++ app/src/main/java/org/sopt/at/data/remote/UserRemoteDataSource.kt
@@ -1,21 +1,10 @@
 package org.sopt.at.data.remote
 
-import org.sopt.at.domain.model.User
 import org.sopt.at.remote.base.BaseResponse
 import org.sopt.at.remote.model.MyNicknameResponse
 import org.sopt.at.remote.model.NicknamesResponse
-import org.sopt.at.remote.model.SignInResponse
 
 interface UserRemoteDataSource {
-
-    suspend fun postSignUp(
-        userInfo: User
-    ): BaseResponse
-
-    suspend fun postSignIn(
-        id: String,
-        password: String
-    ): SignInResponse
 
     suspend fun getMyNickname(): MyNicknameResponse
