+++ app/src/main/java/org/sopt/at/domain/usecase/RequestSignInUseCase.kt
@@ -1,17 +1,17 @@
 package org.sopt.at.domain.usecase
 
-import org.sopt.at.domain.repository.UserRepository
+import org.sopt.at.domain.repository.AuthRepository
 import org.sopt.at.remote.base.BaseResponse
 import javax.inject.Inject
 
 class RequestSignInUseCase @Inject constructor(
-    private val userRepository: UserRepository
+    private val authRepository: AuthRepository
 ) {
     suspend fun invoke(
         id: String,
         password: String,
     ): BaseResponse {
-        return userRepository.requestSignIn(
+        return authRepository.requestSignIn(
             id = id,
             password = password
         )
