+++ app/src/main/java/org/sopt/at/domain/usecase/RequestSignUpUseCase.kt
@@ -1,19 +1,19 @@
 package org.sopt.at.domain.usecase
 
 import org.sopt.at.domain.model.User
-import org.sopt.at.domain.repository.UserRepository
+import org.sopt.at.domain.repository.AuthRepository
 import org.sopt.at.remote.base.BaseResponse
 import javax.inject.Inject
 
 class RequestSignUpUseCase @Inject constructor(
-    private val userRepository: UserRepository
+    private val authRepository: AuthRepository
 ) {
     suspend fun invoke(
         id: String,
         password: String,
         nickname: String
     ): BaseResponse {
-        return userRepository.requestSignUp(
+        return authRepository.requestSignUp(
             User(
                 id = id,
                 password = password,
