+++ app/src/main/java/org/sopt/at/remote/api/AuthApiService.kt
@@ -1,0 +1,26 @@
+package org.sopt.at.remote.api
+
+import org.sopt.at.remote.api.interceptor.AuthType
+import org.sopt.at.remote.model.SignInRequest
+import org.sopt.at.remote.model.SignInResponse
+import org.sopt.at.remote.model.SignUpRequest
+import org.sopt.at.remote.model.SignUpResponse
+import retrofit2.http.Body
+import retrofit2.http.POST
+import retrofit2.http.Tag
+
+interface AuthApiService {
+    // 회원가입
+    @POST("/api/v1/auth/signup")
+    suspend fun postSignUp(
+        @Body signUpRequest: SignUpRequest,
+        @Tag authType: AuthType = AuthType.NO_AUTH
+    ): SignUpResponse
+
+    // 로그인
+    @POST("/api/v1/auth/signin")
+    suspend fun postSignIn(
+        @Body signInRequest: SignInRequest,
+        @Tag authType: AuthType = AuthType.NO_AUTH
+    ): SignInResponse
+}
