+++ app/src/main/java/org/sopt/at/remote/di/RemoteDataSourceModule.kt
@@ -4,13 +4,21 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.sopt.at.data.remote.AuthRemoteDataSource
 import org.sopt.at.data.remote.UserRemoteDataSource
+import org.sopt.at.remote.impl.AuthRemoteDataSourceImpl
 import org.sopt.at.remote.impl.UserRemoteDataSourceImpl
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 internal abstract class RemoteDataSourceModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindAuthRemoteDataSource(
+        source: AuthRemoteDataSourceImpl
+    ): AuthRemoteDataSource
 
     @Binds
     @Singleton
