+++ app/src/main/java/org/sopt/at/remote/impl/UserRemoteDataSourceImpl.kt
@@ -3,68 +3,25 @@ package org.sopt.at.remote.impl
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.sopt.at.data.remote.UserRemoteDataSource
-import org.sopt.at.domain.model.User
 import org.sopt.at.remote.ErrorHandler.handleError
-import org.sopt.at.remote.RemoteMapper.toDTO
-import org.sopt.at.remote.api.ApiService
+import org.sopt.at.remote.api.UserApiService
 import org.sopt.at.remote.base.BaseResponse
 import org.sopt.at.remote.model.MyNicknameResponse
 import org.sopt.at.remote.model.NicknameEditRequest
 import org.sopt.at.remote.model.NicknameResult
 import org.sopt.at.remote.model.NicknamesResponse
 import org.sopt.at.remote.model.NicknamesResult
-import org.sopt.at.remote.model.SignInRequest
-import org.sopt.at.remote.model.SignInResponse
-import org.sopt.at.remote.model.SignInResult
 import timber.log.Timber
 import javax.inject.Inject
 
 class UserRemoteDataSourceImpl @Inject constructor(
-    private val apiService: ApiService
+    private val userApiService: UserApiService
 ): UserRemoteDataSource {
-    override suspend fun postSignUp(userInfo: User): BaseResponse {
-        var response = BaseResponse()
-        withContext(Dispatchers.IO) {
-            runCatching {
-                apiService.postSignUp(
-                    userInfo.toDTO()
-                )
-            }.onSuccess {
-                response = it
-                Timber.d("postSignUp success: $it")
-            }.onFailure { exception ->
-                response = exception.handleError()
-                Timber.d("postSignUp fail: ${response.message}")
-            }
-        }
-        return response
-    }
-
-    override suspend fun postSignIn(
-        id: String,
-        password: String
-    ): SignInResponse {
-        var response = SignInResponse(result = SignInResult(-1))
-        withContext(Dispatchers.IO) {
-            runCatching {
-                apiService.postSignIn(
-                    SignInRequest(id, password)
-                )
-            }.onSuccess {
-                response = it
-                Timber.d("postSignIn success: $it")
-            }.onFailure { exception ->
-                Timber.d("postSignIn fail: $exception")
-            }
-        }
-        return response
-    }
-
     override suspend fun getMyNickname(): MyNicknameResponse {
         var response = MyNicknameResponse(result = NicknameResult(""))
         withContext(Dispatchers.IO) {
             runCatching {
-                apiService.getMyNickname()
+                userApiService.getMyNickname()
             }.onSuccess {
                 response = it
                 Timber.d("getMyNickname success: $it")
@@ -80,7 +37,7 @@ class UserRemoteDataSourceImpl @Inject constructor(
         var response = BaseResponse()
         withContext(Dispatchers.IO) {
             runCatching {
-                apiService.patchNickname(
+                userApiService.patchNickname(
                     NicknameEditRequest(
                         nickname = nickname
                     )
@@ -100,7 +57,7 @@ class UserRemoteDataSourceImpl @Inject constructor(
         var response = NicknamesResponse(result = NicknamesResult(listOf()))
         withContext(Dispatchers.IO) {
             runCatching {
-                apiService.getNicknames(searchNickname)
+                userApiService.getNicknames(searchNickname)
             }.onSuccess {
                 response = it
                 Timber.d("getNicknames success: $it")
