+++ app/src/main/java/org/sopt/at/remote/model/AuthDto.kt
@@ -46,32 +46,3 @@ data class SignInResponse(
 data class SignInResult(
     val userId: Long
 )
-
-/** 닉네임 */
-@Serializable
-data class MyNicknameResponse(
-    @SerialName("data")
-    val result: NicknameResult
-) : BaseResponse()
-
-@Serializable
-data class NicknameResult(
-    val nickname: String
-)
-
-@Serializable
-data class NicknameEditRequest(
-    val nickname: String
-)
-
-/** 유저 닉네임 검색 */
-@Serializable
-data class NicknamesResponse(
-    @SerialName("data")
-    val result: NicknamesResult
-): BaseResponse()
-
-@Serializable
-data class NicknamesResult(
-    val nicknameList: List<String>
-)
